+++ R/pmcmc_utils.R
@@ -135,7 +135,7 @@ sample_pmcmc <- function(pmcmc_results,
 
   trajectories <- array(NA,
                         dim = c(max_rows, ncol(traces[[1]]), length(traces)),
-                        dimnames = list(max_date_names, NULL, NULL))
+                        dimnames = list(max_date_names, colnames(traces[[1]]), NULL))
 
   # fill the tail of the array slice
   # This is so that the end of the trajectories array is populated,